+++ src/Controller/Component/GoogleAuthenticatorComponent.php
@@ -11,11 +11,11 @@
 
 namespace CakeDC\Users\Controller\Component;
 
+use CakeDC\Users\Auth\TwoFactorAuthenticationCheckerInterface;
 use Cake\Controller\Component;
 use Cake\Core\Configure;
 use Cake\Event\Event;
 use Cake\Utility\Security;
-use CakeDC\Users\Auth\TwoFactorAuthenticationCheckerInterface;
 use InvalidArgumentException;
 use RobThree\Auth\TwoFactorAuth;
 
@@ -85,28 +85,5 @@ class GoogleAuthenticatorComponent extends Component
     public function getQRCodeImageAsDataUri($issuer, $secret)
     {
         return $this->tfa->getQRCodeImageAsDataUri($issuer, $secret);
-    }
-
-    /**
-     * Get the two factor authentication checker
-     *
-     * @return TwoFactorAuthenticationCheckerInterface
-     */
-    public function getChecker()
-    {
-        if ($this->checker !== null) {
-            return $this->checker;
-        }
-        $className = Configure::read('GoogleAuthenticator.checker');
-
-        $interfaces = class_implements($className);
-        $required = 'CakeDC\Users\Auth\TwoFactorAuthenticationCheckerInterface';
-
-        if (in_array($required, $interfaces)) {
-            $this->checker = new $className();
-
-            return $this->checker;
-        }
-        throw new InvalidArgumentException("Invalid config for 'GoogleAuthenticator.checker', '$className' does not implement '$required'");
     }
 }
