+++ tests/TestCase/Controller/Component/GoogleAuthenticatorComponentTest.php
@@ -135,28 +135,4 @@ class GoogleAuthenticatorComponentTest extends TestCase
         $verified = $this->Controller->GoogleAuthenticator->verifyCode($secret, $verificationCode);
         $this->assertTrue($verified);
     }
-
-    /**
-     * Test getChecker method
-     *
-     * @return void
-     */
-    public function testGetChecker()
-    {
-        $result = $this->Controller->GoogleAuthenticator->getChecker();
-        $this->assertInstanceOf(DefaultTwoFactorAuthenticationChecker::class, $result);
-    }
-
-    /**
-     * Test getChecker method
-     *
-     * @return void
-     */
-    public function testGetCheckerInvalidInterface()
-    {
-        Configure::write('GoogleAuthenticator.checker', 'stdClass');
-        $this->expectException(\InvalidArgumentException::class);
-        $this->expectExceptionMessage("Invalid config for 'GoogleAuthenticator.checker', 'stdClass' does not implement 'CakeDC\Users\Auth\TwoFactorAuthenticationCheckerInterface'");
-        $this->Controller->GoogleAuthenticator->getChecker();
-    }
 }
